refactor: create username.js module
